start generate process with threads

+++ lab-02/by/katierevinska/docks_and_hobos/Ship.java
@@ -2,9 +2,9 @@ package by.katierevinska.docks_and_hobos;
 
 public class Ship {
     private String cargoType;
-    private int shipCapacity;
+    private Long shipCapacity;
 
-Ship(String cargoType, int capacity){
+Ship(String cargoType, Long capacity){
     this.cargoType= cargoType;
     this.shipCapacity = capacity;
 }
